Extract MCP work proof guidance helpers
Refs #320
